pyproject.toml
[tool.poetry]
name = "sensorqualityclassifier"
version = "0.1.0"
description = ""
authors = ["Krunal Shambharkar <krunalss@outlook.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.9.18"
attrs = "19.3.0"
certifi = "2019.11.28"
Click = "7.1.2"
cycler = "0.10.0"
Flask = "1.1.1"
Flask-Cors = "3.0.8"
importlib-metadata = "1.7.0"
Flask-MonitoringDashboard = "3.0.6"
itsdangerous = "1.1.0"
Jinja2 = "2.11.0"
joblib = "0.14.1"
jsonschema = "3.2.0"
kiwisolver = ">1.3.0"
kneed = "0.5.1" 
MarkupSafe = "1.1.1"
matplotlib = "3.8.2"
more-itertools = "8.1.0"
numpy = "^1.19.0"
pandas = "2.0.3"
pyparsing = "2.4.6"
pyrsistent = "0.15.7"
python-dateutil = "2.8.2"
pytz = ">=2020.1"
PyYAML = "5.3"
regexp = "0.1"
scikit-learn = "^0.24.0"
six = "1.14.0"
sklearn = "0.0"
Werkzeug = "0.16.1"
wincertstore = "0.2"
xgboost = "0.90"
zipp = "2.0.1"
requests = "2.31.0"
hopsworks = "3.4.4"
ipywidgets = "^8.1.2"
python-dotenv = "^1.0.1"
hops = "^3.7.0.0"
seaborn = "^0.13.2"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
